RFCT: refactor function names and extract out for readability.
