+++ backend/src/apis/v2/helpers/image_process_utils.py
@@ -1,16 +1,12 @@
 import cv2
 import numpy as np
 
-from apis.v2.helpers.processor.batch_processor import BatchProcessor
-from apis.v2.helpers.processor.chip_processor import ChipProcessor
 from constants.colors import BGRColors
 from constants.image_thresholds import ImageThreshold
-from db.models.image_settings import ImageSettings
 from schemas.contours import ContourInfo, ContourList
 from utils.image_process.blob_handler import BlobHandler
 from utils.image_process.border_creator import BorderCreator
 from utils.image_process.contour_handler import ContourHandler
-from utils.image_process.mask_handler import MaskHandler
 
 
 def create_border(image: np.ndarray, padding: int = 0, crop_size: int = 0):
@@ -31,29 +27,6 @@ def create_contour_list(mask_image: np.ndarray) -> ContourList:
     return ContourHandler.filter_and_build_contour_info(
         contours, ImageThreshold.DENOISE_THRESHOLD
     )
-
-
-def process_batch(mask_handler: MaskHandler, image_settings: ImageSettings):
-    """Processes the image in batches."""
-    batch_processor = BatchProcessor(
-        mask_handler, image_settings.batch_erode, image_settings.batch_close
-    )
-    batch_processor.get_batch_data()
-    return batch_processor
-
-
-def process_chip(
-    mask_handler: MaskHandler, border_pad: int, image_settings: ImageSettings
-):
-    """Processes the chip data from the mask handler."""
-    chip_processor = ChipProcessor(
-        mask_handler,
-        image_settings.chip_erode,
-        image_settings.chip_close,
-        border_pad,
-        image_settings.crop_size,
-    )
-    return chip_processor
 
 
 def check_single(
